Allow additional whitespace in downstream issue tagging
The current regular expression used to match the downstream issue
line does not allow for the line to be indented with whitespace.
This can be relaxed to prevent the check failing unnecessarily.

Additionally this adds a dryrun option to the script, to prevent
the script making modifications during local testing.