[ATfE] Build AArch64 library variants with `-ffixed-x18` (#485)
In the procedure call standard, `X18` is reserved for the platform. As a
bare-metal toolchain we don't know what the platform will be so avoid
using `X18` in case it is used by other code.

The clang's `-fsanitize=shadow-call-stack` is one such feature that
relies on `X18` not being modified by the library.