[ATfE] Set xfails indirectly instead of modifying test sources (#487)
Certain tests will fail when run with ATfE's compiler and libraries. The
toolchain currently modifies the test sources to change the expected
result, but this requires maintaining downstream patches. The LLVM test
driver, lit, provides options to override the expected result set in the
file through command line and environment variables, which would remove
this burden.

This patch makes use of the `LIT_XFAILS` and `LIT_XFAILS_NOT`
environment variables, which will take precedence over what is set in
the sources.

A new script is provided which will generate a list of tests for a given
project/library variant that are expected to give a different result.
This is largely based on the existing set of downstream patches. This
also has the advantage of allowing xfails to be set for specific
variants, instead of globally. This currently only supports the clang,
compiler-rt and libcxx projects.

In order to prepare the correct environment for lit, a wrapper script is
also provided which will accept files containing test lists to be loaded
into the relevant environment variable. Since the test lists can be
quite long, this prevents the command line getting significantly
bloated.