Temporarily disable automerge for 21.x release branch (#494)
This temporarily disables Automerge workflow runs for the 21.x release
branch. As the 21.x release approaches upstream, this is necessary to
ensure the commit used for the Arm Toolchain release is consistent
with the one used upstream.

While the Automerge is disabled for the branch, any incoming changes
from upstream will be reviewed and manually merged into the downstream
release/arm-software/21.x branch.

The workflow will be re-enabled for the branch once the release is done.