[ATfE] Add AArch64 samples for LLVM-libc (#496)
This is based on the AArch64 sample from picolibc. This requires a minor
hack. We need to place the _start at the beginning of the boot_flash
section, and run QEMU from 0x0. This allows QEMU to turn off
multithreading, and reach the _start symbol.

We use the section `.text.init.enter` for this. We do need extra space
in `boot_flash`, because we also put the page table in this section.