[multilib-generate.py] Fix parsing of AArch32 extensions (#511)
The function `get_extension_list()` in this script parses the output of
`clang --target=[...] --print-supported-extensions` to make a list of
architecture extension names. It's supposed to work for both AArch32 and
AArch64, but in fact it was ignoring all AArch32 extensions, because
clang gives differently formatted output for the two architectures, the
parsing code was specific to the AArch64 version, and it silently
ignored any unrecognised line instead of reporting an error.

I've replaced the previous parser with one that can handle both formats.
Also, it reads the output more thoroughly, and objects to anything it
doesn't completely understand. So if clang changes the format of this
output in future, we'll have to fix this script, but we'll at least know
that, and won't find out that it's been quietly doing the wrong thing.

The new code is complicated enough that I've included a couple of unit
tests, which you can run with the command
```
python3 -m unittest arm-software/embedded/arm-multilib/multilib-generate.py
```
But I haven't made that test run automatically; it's there to be run
manually the next time someone edits the affected code. (A more
important test is to ensure the script continues to be happy with
clang's actual output, rather than the hardcoded example output in the
unit tests.)

Fixes #484, I hope.