[ATfE] Run picolibc tests via lit (#523)
The picolibc project is built and tested through meson, which differs
from the other projects which use CMake to build and lit to run the
tests. It would be useful to run all the tests using the same
infrastructure, so that they can all be controlled in the same way. In
particular, when tests are expected to fail, meson does not provide a
mechanism to handle this, whereas lit does.

This patch allows the picolibc test suite to be executed through lit, by
generating a lit test for every meson test. The existing lit controls
for expected failures can then be used when running the generated test
suite.

The overriding of return codes has been removed from the picolibc test
wrapper, and instead added to the new list of expected failures. Only
the FVP test failures were still relevant: the QEMU failures referenced
variants that no longer exist, and the tests themselves actually pass
now. Since the tests were treated as skipped, this has gone unnoticed.
In future, if the state of a test changes, lit will flag that the result
was unexpected.

A CMake target has been retained to optionally run the picolibc tests
through meson for debugging purposes, since lit's interpretation of the
results will differ from the unmodified results from meson.