[ATfE] Use a consistent name for xml results (#536)
Currently, some test suites are hard-coded to use a specific name/path
to write xml results to. Others rely on this being set manually. Some of
the CI scripts don't set it at all for certain test targets.

This patch aligns all the test suites to use the same name, controlled
via the same mechanism. A user or CI system can set the
`--xunit-xml-output` lit option via the `LIT_OPTS` environment variable
to control the name of xml file, and therefore know the correct name to
look for when storing the results.

A few minor corrections have also been made to the CI scripts for
running tests:

* Since ninja defaults to building in parallel, there is no reason to
set the `-j` option.
* `check-all` was getting built twice in some scripts.
* Comments and lit options updated to reflect that picolibc testing does
in fact use lit now.