[ATfL] Make sure rpath is added to the helper programs used by the lit tests
One of the most recent changes in the testing routines resulted in the
LD_LIBRARY_PATH env variable not being passed to the executed tests.

As we are using libc++ library built during one of the earlier stages,
we were relying on the ability to find it on the LD_LIBRARY_PATH list.

This patch creates a config files for the bootstrap compiler (which is
being used for compiling helper programs executed by the lit tests)
which makes it obligatory to pass -rpath=... flag to the linker,
so the path to precompiled libc++ library is being set as an runtime
path. After the round of tests, these config files are being deleted,
so they should not infect the subsequent stages.