[ATfE] Create custom timeout mechanism for picolibc test runs
picolibc defines timeout values for each test. In general these are
enough for our test runs, however some variants can be slower due to
less intensive code optimizations. One example are the Armv8.1-M Main
PACBTI variants optimized for code size.

For such variants, we need a way to specify timeouts values different
than the ones set by the picolibc tests definitions.

This patch introduces a mechanism to set timeout multipliers for library
variants. A new json file is created in order to contain the mapping
from library variants to timeout multipliers. These multipliers are used
in the `gen-picolibc-lit-tests.py` script to generate the `meson test`
command line to be used in the test runs. `meson test` takes the `-t`
argument, which is the timeout multiplier.

As the first use, the timeouts json file contains a multiplier for the
M-profile PACBTI variants for code size.