[Downstream change] Cherry-pick of [TargetParser] Parse Amazon Linux triple and handle RuntimeLibcalls (#546)
Original PR under review: llvm/llvm-project#136114

As the commit introduced by #535 has resulted in a faulty C++ compiler
which could not see system-provided standard C++ library, an alternative
solution is now being cherry-picked here which appears to work with
Amazon Linux 2023.

This is an one off change for the ATfL22 release.

Downstream issue: #533

Co-authored-by: Benjamin Maxwell <benjamin.maxwell@arm.com>