[ATfE] Create custom timeout mechanism for picolibc test runs (#548)
picolibc defines timeout values for each test. In general these are
enough for our test runs, however some variants can be slower due to
less intensive code optimizations or security features. One example are
the Armv8.1-M Main PACBTI variants optimized for code size.

For such variants, we need a way to specify timeouts values different
than the ones set by the picolibc tests definitions.

This patch introduces a mechanism to set timeout multipliers for library
variants. In the arm-runtimes' `CMakeLists.txt`, there's a new section
where one can specify conditions based on library build options to
determine different timeout multipliers. These multipliers are then used
in the `gen-picolibc-lit-tests.py` script to generate the `meson test`
command line to be used in the test runs. `meson test` takes the `-t`
argument, which is the timeout multiplier.

For a limited number of custom timeout conditions, this approach scales
without problems. But if in the future the list of custom conditions
grows, it might be worth extracting the logic out of `CMakeLists.txt`
into a new file in order to have a separation between build system's
code and data.