[ATfE] Update gettimeofday xfails for new FVP version (#584)
The semi-hosting implementation of the SYS_TICKFREQ call was broken in
version 11.27 of the FVPs used in testing, and so the picolibc test
semihost-gettimeofday needed to be xfailed. This was fixed in version
11.28 of the models, which is the version fetched by the get_fvps.sh
script for the AEMv8A and AEMv8R models. For variants using these
models, the xfail can be removed. An updated 11.28 version of the
Corstone-310 model is not yet available, so the variants using this
model still require the xfail.