[ATfE] Remove lit wrapper script
This passes the xfail lists to lit using argparse's fromfile_prefix
character, which is set to `@` in lit already. This removes the need
to modify lit with additional options, and maintain a wrapper script.

Additionally, the clang xfail list has been made more specific to
clang, so that only check-clang requires the additional list. This
prevents every LLVM project requiring the additional lit arguments
for their check target.