[ATfE] Fix invalid register error `ELR_hyp` in exception handling on Armv7-A
The exception handling functions print out the value of the link
register for logging purposes. The problem with this is in reading
the `ELR_hyp` register.

This register is only accessible if the processor is in Hypervisor mode.
This mode is always available on Armv8-A, but it is optional on Armv7-A,
conditioned on the presence of the Virtualization Extension. Attempts to
read/write to this register on the latter, if the Virtualization
Extension is not in, cause a compiler error, and possibly a fault at
runtime if it hadn't been caught by the compiler.

This patch fixes this by reading from the correct link register
depending on the configuration. If the processor is not in Hypervisor
mode, the code must read from `LR` instead of `ELR_hyp`.