[ATfE] Fix invalid register error `ELR_hyp` in exception handling on Armv7-A (#614)
The exception handling functions print out the value of the link
register for logging purposes. The problem with this is in reading the
`ELR_hyp` register.

This register is only accessible if the processor is in Hypervisor mode.
This mode is always available on Armv8-A, but it is optional on Armv7-A,
conditioned on the presence of the Virtualization Extension. Attempts to
read/write to this register on the latter, if the Virtualization
Extension is not in, cause a compiler error, and possibly a fault at
runtime if it hadn't been caught by the compiler.

This patch fixes this by reading from the correct link register
depending on the configuration. If the processor is not in Hypervisor
mode, the code must read from `LR` instead of `ELR_hyp`.

In one of the handlers, the instruction pointed by the link register is
printed out. In Thumb mode, the link register might contain an address
that is not 4-byte aligned, therefore the load that retrieves the
instruction is now done 2 bytes at a time to avoid misaligned access. We
could have code that selectively loads either 4 bytes at once or 2 times
2 bytes depending on the ISA state, but it doesn't seem worth it.