[ATfE] Extend xfail mechanism to exclude tests, and exclude test-hello-raw
The test-hello-raw picolibc test expects qemu to send data on the
serial port, which the downstream test runner scripts do not
currently support. While this test can be set to be expected to fail,
the test will wait the full timeout duration so it would be better to
skip it until the test runner can be updated.

The lit test  driver already supports removing tests from the available
set through the --filter-out option, so this patch extends the existing
xfail mechanism to also allow for excluding tests through this argument.
Unlike the xfail/xfail-not options, this expects a regular expression,
but simply or'ing the filenames is sufficient.

Since skipping the test reduces test coverage, the --show-excluded
option is also enabled by default so that it is clear in the logs when
a test is removed.